Revert "L515 motion correction 2"
