+++ src/ds5/ds5-private.h
@@ -561,8 +561,7 @@ namespace librealsense
         enum imu_eeprom_id : uint16_t
         {
             dm_v2_eeprom_id     = 0x0101,   // The pack alignment is Big-endian
-            tm1_eeprom_id       = 0x0002,
-            l500_eeprom_id      = 0x0105
+            tm1_eeprom_id       = 0x0002
         };
 
         struct depth_table_control
