+++ src/l500/l500-motion.cpp
@@ -93,26 +93,6 @@ namespace librealsense
         const l500_motion* _owner;
     };
 
-    std::vector<uint8_t> l500_motion::get_imu_eeprom_raw() const
-    {
-        // read imu calibration table on L515
-        // READ_TABLE 0x243 0
-        command cmd(ivcam2::READ_TABLE, 0x243, 0);
-
-        std::vector<uint8_t> res;
-
-        try
-        {
-            res = _hw_monitor->send(cmd);
-        }
-        catch (std::exception &e) {
-            LOG_WARNING(e.what());
-        }
-        catch (...) {}
-
-        return res;
-    }
-
     std::shared_ptr<synthetic_sensor> l500_motion::create_hid_device(std::shared_ptr<context> ctx, const std::vector<platform::hid_device_info>& all_hid_infos)
     {
         if (all_hid_infos.empty())
@@ -137,31 +117,16 @@ namespace librealsense
         hid_ep->get_option(RS2_OPTION_GLOBAL_TIME_ENABLED).set(0);
         hid_ep->register_option(RS2_OPTION_GLOBAL_TIME_ENABLED, enable_global_time_option);
 
-        // register pre-processing
-        bool enable_imu_correction = false;
-        std::shared_ptr<enable_motion_correction> mm_correct_opt = nullptr;
-
-        //  Motion intrinsic calibration presents is a prerequisite for motion correction.
-        try
-        {
-            if (_mm_calib)
-            {
-                mm_correct_opt = std::make_shared<enable_motion_correction>(hid_ep.get(), option_range{ 0, 1, 1, 1 });
-                hid_ep->register_option(RS2_OPTION_ENABLE_MOTION_CORRECTION, mm_correct_opt);
-            }
-        }
-        catch (...) {}
-
         hid_ep->register_processing_block(
             { {RS2_FORMAT_MOTION_XYZ32F, RS2_STREAM_ACCEL} },
             { {RS2_FORMAT_MOTION_XYZ32F, RS2_STREAM_ACCEL} },
-            [&, mm_correct_opt]() { return std::make_shared<acceleration_transform>(_mm_calib, mm_correct_opt); }
+            []() { return std::make_shared<acceleration_transform>(); }
         );
 
         hid_ep->register_processing_block(
             { {RS2_FORMAT_MOTION_XYZ32F, RS2_STREAM_GYRO} },
             { {RS2_FORMAT_MOTION_XYZ32F, RS2_STREAM_GYRO} },
-            [&, mm_correct_opt]() { return std::make_shared<gyroscope_transform>(_mm_calib, mm_correct_opt); }
+            []() { return std::make_shared<gyroscope_transform>(); }
         );
 
         return hid_ep;
@@ -172,24 +137,6 @@ namespace librealsense
           _accel_stream(new stream(RS2_STREAM_ACCEL)),
          _gyro_stream(new stream(RS2_STREAM_GYRO))
     {
-        _imu_eeprom_raw = [this]() { return get_imu_eeprom_raw(); };
-
-        if (!_imu_eeprom_raw->empty())
-        {
-            _mm_calib = std::make_shared<mm_calib_handler>(*_imu_eeprom_raw);
-
-            _accel_intrinsic = std::make_shared<lazy<ds::imu_intrinsic>>([this]() { return _mm_calib->get_intrinsic(RS2_STREAM_ACCEL); });
-            _gyro_intrinsic = std::make_shared<lazy<ds::imu_intrinsic>>([this]() { return _mm_calib->get_intrinsic(RS2_STREAM_GYRO); });
-            // use predefined values extrinsics
-            _depth_to_imu = std::make_shared<lazy<rs2_extrinsics>>([this]() { return _mm_calib->get_extrinsic(RS2_STREAM_ACCEL); });
-        }
-
-        // Make sure all MM streams are positioned with the same extrinsics
-        environment::get_instance().get_extrinsics_graph().register_extrinsics(*_depth_stream, *_accel_stream, _depth_to_imu);
-        environment::get_instance().get_extrinsics_graph().register_same_extrinsics(*_accel_stream, *_gyro_stream);
-        register_stream_to_extrinsic_group(*_gyro_stream, 0);
-        register_stream_to_extrinsic_group(*_accel_stream, 0);
-
         auto hid_ep = create_hid_device(ctx, group.hid_devices);
         if (hid_ep)
         {
